refactor(Layer): add c3DTilesLayer.
