+++ examples/3dtiles_basic.html
@@ -45,25 +45,22 @@
 
             // Create a new Layer 3d-tiles For DiscreteLOD
             // -------------------------------------------
-            var $3dTilesLayerDiscreteLOD = new itowns.GeometryLayer('3d-tiles-discrete-lod', new itowns.THREE.Group());
-
-            $3dTilesLayerDiscreteLOD.name = 'DiscreteLOD';
-            $3dTilesLayerDiscreteLOD.url = 'https://raw.githubusercontent.com/AnalyticalGraphicsInc/3d-tiles-samples/master/tilesets/TilesetWithDiscreteLOD/tileset.json';
-            $3dTilesLayerDiscreteLOD.protocol = '3d-tiles'
-            $3dTilesLayerDiscreteLOD.overrideMaterials = true;  // custom cesium shaders are not functional
+            var $3dTilesLayerDiscreteLOD = new itowns.C3DTilesLayer('3d-tiles-discrete-lod', {
+                name: 'DiscreteLOD',
+                url: 'https://raw.githubusercontent.com/AnalyticalGraphicsInc/3d-tiles-samples/master/tilesets/TilesetWithDiscreteLOD/tileset.json',
+            }, view);
 
             itowns.View.prototype.addLayer.call(view, $3dTilesLayerDiscreteLOD);
 
             // Create a new Layer 3d-tiles For Viewer Request Volume
             // -----------------------------------------------------
 
-            var $3dTilesLayerRequestVolume = new itowns.GeometryLayer('3d-tiles-request-volume', new itowns.THREE.Group());
+            var $3dTilesLayerRequestVolume = new itowns.C3DTilesLayer('3d-tiles-request-volume', {
+                name: 'RequestVolume',
+                url: 'https://raw.githubusercontent.com/AnalyticalGraphicsInc/3d-tiles-samples/master/tilesets/TilesetWithRequestVolume/tileset.json',
+                sseThreshold: 1,
+            }, view);
 
-            $3dTilesLayerRequestVolume.name = 'RequestVolume';
-            $3dTilesLayerRequestVolume.url = 'https://raw.githubusercontent.com/AnalyticalGraphicsInc/3d-tiles-samples/master/tilesets/TilesetWithRequestVolume/tileset.json';
-            $3dTilesLayerRequestVolume.protocol = '3d-tiles'
-            $3dTilesLayerRequestVolume.overrideMaterials = true;  // custom cesium shaders are not functional
-            $3dTilesLayerRequestVolume.sseThreshold = 1;
             // add an event for picking the 3D Tiles layer and displaying
             // information about the picked feature in an html div
             var pickingArgs = {};
@@ -72,8 +69,9 @@
             pickingArgs.layer = $3dTilesLayerRequestVolume;
             itowns.View.prototype.addLayer.call(view,
                 $3dTilesLayerRequestVolume).then(function _() {
-                    window.addEventListener('mousemove', fillHTMLWithPickingInfo.bind(pickingArgs),
-                        false); });
+                    window.addEventListener('mousemove',
+                        (event) => fillHTMLWithPickingInfo(event, view, pickingArgs),false);
+                });
 
             // Add the UI Debug
             var d = new debug.Debug(view, menuGlobe.gui);
