+++ examples/3dtiles_batch_table.html
@@ -32,7 +32,6 @@
                 heading: -145
             };
 
-
             var viewerDiv = document.getElementById('viewerDiv');
 
             var view = new itowns.GlobeView(viewerDiv, placement);
@@ -50,19 +49,16 @@
                         .then(menuGlobe.addLayerGUI.bind(menuGlobe));
                 });
 
-            // Create a new 3D tiles layer with batch table hierarchy extension
-            var $3dTilesLayerBTHierarchy = new itowns.GeometryLayer(
-                '3d-tiles-bt-hierarchy', new itowns.THREE.Group());
-            $3dTilesLayerBTHierarchy.name = 'BTHierarchy';
-            $3dTilesLayerBTHierarchy.url =
-                'https://raw.githubusercontent.com/AnalyticalGraphicsInc/cesium/master/Apps/SampleData/Cesium3DTiles/Hierarchy/BatchTableHierarchy/tileset.json';
-            $3dTilesLayerBTHierarchy.protocol = '3d-tiles';
-            $3dTilesLayerBTHierarchy.overrideMaterials = true;
-
             // Create and define 3D Tiles extensions needed in this example
             itowns.$3dTilesExtensions.registerExtension(
                 "3DTILES_batch_table_hierarchy",
                 itowns.BatchTableHierarchyExtensionParser.parse);
+            // Create a new 3D tiles layer with batch table hierarchy extension
+            var $3dTilesLayerBTHierarchy = new itowns.C3DTilesLayer(
+                '3d-tiles-bt-hierarchy', {
+                    name: 'BTHierarchy',
+                    url: 'https://raw.githubusercontent.com/AnalyticalGraphicsInc/cesium/master/Apps/SampleData/Cesium3DTiles/Hierarchy/BatchTableHierarchy/tileset.json',
+            }, view);
 
             // add an event for picking the 3D Tiles layer and displaying
             // information about the picked feature in an html div
@@ -70,10 +66,10 @@
             pickingArgs.htmlDiv = document.getElementById('featureInfo');
             pickingArgs.view = view;
             pickingArgs.layer = $3dTilesLayerBTHierarchy;
-            itowns.View.prototype.addLayer.call(view,
-                $3dTilesLayerBTHierarchy).then(function _() {
-                window.addEventListener('mousemove', fillHTMLWithPickingInfo.bind(pickingArgs),
-                    false); });
+            itowns.View.prototype.addLayer.call(view, $3dTilesLayerBTHierarchy).then(function _() {
+                window.addEventListener('mousemove',
+                    (event) => fillHTMLWithPickingInfo(event, view, pickingArgs),false);
+            });
 
             // Add a debug UI
             var d = new debug.Debug(view, menuGlobe.gui);
